integrated redis

+++ src/main/java/com/assignment/notificationservice/controller/BlacklistedNumbersController.java
@@ -5,11 +5,16 @@ import com.assignment.notificationservice.dto.BlacklistedNumbersDTO;
 import com.assignment.notificationservice.dto.BlacklistedNumbersResponseDTO;
 import com.assignment.notificationservice.dto.GeneralMessageDTO;
 import com.assignment.notificationservice.entity.BlacklistedNumbers;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cache.annotation.EnableCaching;
+import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.NoSuchElementException;
 
@@ -18,28 +23,50 @@ import java.util.NoSuchElementException;
 @EnableCaching
 public class BlacklistedNumbersController {
     BlacklistedNumbersRepository blacklistedNumbersRepository;
+    RedisTemplate redisTemplate;
 
-    @Autowired
-    public BlacklistedNumbersController(BlacklistedNumbersRepository blacklistedNumbersRepository){
-        this.blacklistedNumbersRepository=blacklistedNumbersRepository;
+    private static final Logger LOGGER = LoggerFactory.getLogger(BlacklistedNumbersController.class);
+
+
+    public BlacklistedNumbersController(BlacklistedNumbersRepository blacklistedNumbersRepository, RedisTemplate redisTemplate) {
+        this.blacklistedNumbersRepository = blacklistedNumbersRepository;
+        this.redisTemplate = redisTemplate;
     }
 
+    private static final String KEY="BLACKLISTED_NUMBER";
     @GetMapping("/")
     public ResponseEntity<BlacklistedNumbersResponseDTO> findAllBlacklistedNumbers(){
-        List<String> result=blacklistedNumbersRepository.getAllBlacklistedPhoneNumbers();
-        BlacklistedNumbersResponseDTO responseEntity=new BlacklistedNumbersResponseDTO(result);
-        return ResponseEntity.ok(responseEntity);
+        try {
+            List<String> result=redisTemplate.opsForHash().values(KEY);
+            return ResponseEntity.ok(new BlacklistedNumbersResponseDTO(result));
+        }catch (Exception e){
+            e.printStackTrace();
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new BlacklistedNumbersResponseDTO(new ArrayList<>()));
+        }
     }
 
     @PostMapping("/")
     public ResponseEntity<GeneralMessageDTO> addNewNumbersToBlacklistedNumbers(@RequestBody BlacklistedNumbersDTO blacklistedNumbersDTO){
         var phoneNumbers= blacklistedNumbersDTO.getPhoneNumbers();
-        System.out.println("here are your phone numbers"+ phoneNumbers);
+//        System.out.println("here are your phone numbers dude "+ phoneNumbers);
+        LOGGER.info(String.format("The list of numbers recieved to be blacklisted : %s",phoneNumbers.toString()));
+
         for(var phoneNumber:phoneNumbers){
-            // to do: before saving in the database make sure it isn't already present in the database by making a look up in redis.
-            var blacklistedNumber=new BlacklistedNumbers(phoneNumber);
-            blacklistedNumbersRepository.save(blacklistedNumber);
-            //to do: after saving in the database, save it in the redis as well.
+            try{
+                if(redisTemplate.opsForHash().hasKey(KEY,phoneNumber)) {
+//                    System.out.println("this phone number already exists bro "+phoneNumber);
+                    LOGGER.info("Phone number %s is already blocked.",phoneNumber);
+                    continue;
+                }
+                redisTemplate.opsForHash().put(KEY,phoneNumber,phoneNumber);
+                LOGGER.info("Added phone number %s to redis blacklisted list.",phoneNumber);
+                var blacklistedNumber=new BlacklistedNumbers(phoneNumber);
+                blacklistedNumbersRepository.save(blacklistedNumber);
+                LOGGER.info("Added phone number %s to blacklisted database.",phoneNumber);
+            }catch (Exception e){
+                e.printStackTrace();
+                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new GeneralMessageDTO("false"));
+            }
         }
         return ResponseEntity.ok(new GeneralMessageDTO("Successfully blacklisted"));
     }
@@ -49,9 +76,19 @@ public class BlacklistedNumbersController {
         var phoneNumbers= blacklistedNumbersDTO.getPhoneNumbers();
         System.out.println("here are your phone numbers"+ phoneNumbers);
         for(var phoneNumber:phoneNumbers){
-            // to do: before saving in the database make sure it is already present in the database by making a look up in redis.
-            blacklistedNumbersRepository.deleteSpecificPhoneNumber(phoneNumber);
-            //to do: after deleting from the database, delete it from the redis as well.
+            try{
+                if(!redisTemplate.opsForHash().hasKey(KEY,phoneNumber)){
+                    LOGGER.info(String.format("Phone number %s doesn't exist in the database.",phoneNumber));
+                    continue;
+                }
+                redisTemplate.opsForHash().delete(KEY,phoneNumber);
+                LOGGER.info(String.format("Deleted phone number %s from redis.",phoneNumber));
+                blacklistedNumbersRepository.deleteSpecificPhoneNumber(phoneNumber);
+                LOGGER.info(String.format("Deleted phone number %s from databse.",phoneNumber));
+            }catch(Exception e){
+                e.printStackTrace();
+                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new GeneralMessageDTO("bad request"));
+            }
         }
         return ResponseEntity.ok(new GeneralMessageDTO("Successfully deleted"));
     }
